+++ app/scripts/lib/lforms-fhir-sdc-converter.js
@@ -66,6 +66,10 @@ if (typeof LForms.FHIR_SDC === 'undefined')
       lfData.code = code.code;
       lfData.codeSystem = code.system;
     }
+
+    if(questionnaire.id) {
+      lfData.id = questionnaire.id;
+    }
   }
 
 
@@ -92,6 +96,7 @@ if (typeof LForms.FHIR_SDC === 'undefined')
     _processCodingInstructions(targetItem, qItem);
     _processUnitList(targetItem, qItem);
     _processDefaultAnswer(targetItem, qItem);
+    _processExternallyDefined(targetItem, qItem);
     _processAnswers(targetItem, qItem);
     _processSkipLogic(targetItem, qItem, qResource);
 
@@ -162,6 +167,18 @@ if (typeof LForms.FHIR_SDC === 'undefined')
     }
   }
 
+
+  /**
+   *
+   * @param lfItem
+   * @param qItem
+   * @private
+   */
+  function _processExternallyDefined(lfItem, qItem) {
+    if(qItem.options) {
+      lfItem.externallyDefined = qItem.options.reference;
+    }
+  }
 
   /**
    * Parse questionnaire item for answers list